requirements.txt
matplotlib==3.7.1
numpy==1.24.1
PyYAML==6.0.1
scipy==1.10.1
torch==2.1.0
torchvision==0.16.0
